Change back to one rsa key per machine model.
It doesn't seem possible to have a single ssh/config file that works
across machines so that github will be accessed with the local key. Even
if the public key is shared, the private key has to be transfered
between machines. In order to do the transfer, a one-time use ssh key
would need to be set up anyway, so it doesn't save any effort.